requirements.txt
numpy
opencv-python
Pillow
gradio
pandas
ultralytics
